requirements.txt
groq
fastapi
uvicorn
sentence-transformers
faiss-cpu
python-dotenv
pdfminer.six
tqdm
